1 to N

+++ F. Digits Summation/index.cpp
@@ -1,16 +1,13 @@
-#include<iostream>
+#include <iostream>
 using namespace std;
-int DigitsSum(int a,int b){
-    return a%10+b%10;
+int DigitsSum(long long a, long long b)
+{
+    return a % 10 + b % 10;
 }
 
-
-
-int main(){
-    int a,b;
-    cin>>a>>b;
-    cout<<DigitsSum(a,b);
-
-
-
+int main()
+{
+    long long a, b;
+    cin >> a >> b;
+    cout << DigitsSum(a, b);
 }
